replace fs_root with absolute paths, and rename multifs
