+++ src/test/unit_tests/sudo_index.js
@@ -22,5 +22,5 @@ require('./test_ns_list_objects');
 require('./test_nsfs_access');
 require('./test_bucketspace');
 require('./test_bucketspace_versioning');
-require('./test_bucketspace_multi_fs');
+require('./test_bucketspace_fs');
